requirements.txt
streamlit
pandas
networkx
matplotlib
requests
openrouteservice
python-dotenv